### 2025-11-06 #9378 @editorialbot
[PRE REVIEW]: mapme.biodiversity: Efficient Monitoring of Global Biodiversity Portfolios
**Submitting author:** @goergen95 (<a href="http://orcid.org/0009-0008-5503-7704">Darius A. Görgen</a>)
**Repository:** https://github.com/mapme-initiative/mapme.biodiversity
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.9.5
