+++ dev/pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   quickstart: quickstart.html
   terminology: terminology.html
   workflow: workflow.html
-last_built: 2025-11-06T04:20Z
+last_built: 2025-11-07T04:15Z
 urls:
   reference: https://mapme-initiative.github.io/mapme.biodiversity/reference
   article: https://mapme-initiative.github.io/mapme.biodiversity/articles
