+++ paper/paper.md
@@ -78,93 +78,46 @@ to derive standardized, area-based indicators at scale, supports
 reproducible workflows in \texttt{R}, and facilitates integration with
 other tools via export to standard spatial formats.
 
-Originally developed for conservation monitoring, the package has also
-found applications in adjacent domains such as agriculture, rural
-development, food security, and infrastructure planning. It offers a
-transparent and auditable alternative to proprietary
-platform-as-a-service solutions like Google Earth Engine or
-Microsoft Planetary Computer, which may be inaccessible to users in
-low-connectivity settings, non-academic institutions, or those requiring
-long-term reproducibility not guaranteed by commercial platforms.
-
 # Applications
 
-As an open-source package, the full range of \texttt{mapme.biodiversity}
-applications is difficult to track. However, download statistics
-indicate substantial usage, with 19,599 downloads from the Comprehensive 
-R Archive Network (CRAN) between April 2022 and September 2025. 
 The package is actively used by several institutions for both operational 
-and research purposes.
-
-At the Kreditanstalt für Wiederaufbau (KfW) and the Agence Française de
-Développement -- the German and French public development banks
-for international development -- it is used for internal impact
-evaluations and reporting on funded conservation and development
-programs. At the French National Research Institute for Sustainable
-Development (IRD), it is used in research on policy evaluation and in
-projects aimed at strengthening research capacity in the Global South,
-including ongoing work in Madagascar and Senegal.
+and research purposes. At the Kreditanstalt für Wiederaufbau (KfW) and 
+the Agence Française de Développement -- the German and French public 
+development banks for international development -- it is used for 
+internal impact evaluations and reporting on funded conservation and 
+development programs. At the French National Research Institute for 
+Sustainable Development (IRD), it is used in research on policy evaluation 
+and in projects aimed at strengthening research capacity in the Global 
+South, including ongoing work in Madagascar and Senegal.
 
 # Comparison with other software packages
 
-Different software tools address related needs in conservation
-monitoring and biodiversity assessment. The Digital Observatory for
-Protected Areas provides a centralized, server-based
-system developed by the European Commission to compute global
+The Digital Observatory for Protected Areas provides a centralized, 
+server-based system developed by the European Commission to compute global
 protected-area indicators through standardized, automated workflows
-[@juffe-bignoli_delivering_2024]. In contrast, mapme.biodiversity offers
+[@juffe-bignoli_delivering_2024]. In contrast, \texttt{mapme.biodiversity} offers
 an R-native, decentralized framework that enables users to reproduce
 similar area-based analyses locally, adapt them to specific contexts,
-and extend them with additional datasets. Both share the goal of
-consistent and transparent indicator production, but differ in scale,
-accessibility, and infrastructure dependency.
+and extend them with additional datasets.
 
-The \texttt{wdpar} \texttt{R} package facilitates access to and cleaning
-of spatial data from the World Database on Protected Areas
-[@hanson_wdpar_2022]. \texttt{mapme.biodiversity} is commonly used
-alongside \texttt{wdpar} to compute diverse environmental indicators
-from multiple sources on these cleaned protected area boundaries.
-
-The Global Forest Watch (\texttt{GFW}) API, developed by the World
-Resources Institute, provides access to selected global forest
-monitoring datasets such as tree cover, loss, gain, biomass, and fire
-activity derived from satellite products like Hansen et al.
-[-@hansen_high-resolution_2013]. It allows users to delegate
-computations to a remote cloud infrastructure on these datasets and
-retrieve aggregated statistics for defined areas.
-\texttt{mapme.biodiversity} instead performs all processing locally or
[…]
