+++ dev/reference/check_available_years.md
@@ -1,0 +1,24 @@
+# Helper to check yearly availability
+
+Use this function to check if a specified vector of years intersects
+with the yearly availability of a resource.
+
+## Usage
+
+``` r
+check_available_years(target_years, available_years, indicator)
+```
+
+## Arguments
+
+- target_years:
+
+  Numeric indicating the target year.
+
+- available_years:
+
+  Numeric indicating the available years.
+
+- indicator:
+
+  A character vector with target resource/indicator name.
