+++ dev/pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   quickstart: quickstart.html
   terminology: terminology.html
   workflow: workflow.html
-last_built: 2025-11-09T04:15Z
+last_built: 2025-11-10T04:20Z
 urls:
   reference: https://mapme-initiative.github.io/mapme.biodiversity/reference
   article: https://mapme-initiative.github.io/mapme.biodiversity/articles
