### 2026-01-07 8fd1212
update nelson_et_al
