+++ dev/authors.md
@@ -24,13 +24,13 @@ Source:
 [`DESCRIPTION`](https://github.com/mapme-initiative/mapme.biodiversity/blob/main/DESCRIPTION)
 
 Görgen D, Bhandari O (2026). *mapme.biodiversity: Efficient Monitoring
-of Global Biodiversity Portfolios*. R package version 0.9.5.9000,
+of Global Biodiversity Portfolios*. R package version 0.9.5.9001,
 <https://mapme-initiative.github.io/mapme.biodiversity/>.
 
     @Manual{,
       title = {mapme.biodiversity: Efficient Monitoring of Global Biodiversity Portfolios},
       author = {Darius A. Görgen and Om Prakash Bhandari},
       year = {2026},
-      note = {R package version 0.9.5.9000},
+      note = {R package version 0.9.5.9001},
       url = {https://mapme-initiative.github.io/mapme.biodiversity/},
     }
