+++ dev/reference/humanfootprint_resource.md
@@ -3,7 +3,7 @@
 This resource is part of the publication by Mu et al. (2022) "A global
 record of annual terrestrial Human Footprint dataset from 2000 to 2018".
 It is calculated based on 8 variables representing human pressures on
-natural ecosystems collected at a yearly cadence between 2000 and 2022
+natural ecosystems collected at a yearly cadence between 2000 and 2024
 sampled at a 1km spatial resolution. The variables are used are the
 expansion of built environments (expressed as percentage of built-up
 areas within a grid cell), population density (aggregated at the grid
@@ -17,7 +17,7 @@ theoretical value of the highest human pressure.
 ## Usage
 
 ``` r
-get_humanfootprint(years = 2000:2022)
+get_humanfootprint(years = 2000:2024)
 ```
 
 ## Source
@@ -29,7 +29,7 @@ get_humanfootprint(years = 2000:2022)
 - years:
 
   A numeric vector indicating the years for which to download the human
-  footprint data, defaults to `2000:2022`.
+  footprint data, defaults to `2000:2024`.
 
 ## Value
 
