+++ dev/news/index.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## mapme.biodiversity 0.9.5.9001
+
+### General
+
+- modified
+  [`get_nelson_et_al()`](https://mapme-initiative.github.io/mapme.biodiversity/dev/reference/nelson_et_al.md)
+  to cope with Figshare asynchronous download mechanism
+  ([474](https://github.com/mapme-initiative/mapme.biodiversity/issues/474))
+- updated
+  [`get_humanfootprint()`](https://mapme-initiative.github.io/mapme.biodiversity/dev/reference/humanfootprint_resource.md)
+  to accept years up to 2024
+
 ## mapme.biodiversity 0.9.5.9000
 
 ### General
