### 2026-01-20 bb49491
GDAL and MCP upodates
