### 2026-01-26 #9378 @jatkinson1000
Hi @goergen95 I will be editing this paper with you.

Thank you for your patience over the holiday period and whilst JOSS worked down its backlog.

As you will see from Kristen's comment above, JOSS recently updated its scope requiring a little more information in papers.
Please could you review the updated guidance, make any changes as necessary to your paper, and let me know when you have done so.
After this I'll be able to start looking for reviewers and get the review process started.
