+++ dev/reference/mapme.md
@@ -3,8 +3,10 @@
 `mapme_options()` sets default options for mapme.biodiversity to control
 the behaviour of downstream functions. Mainly, the output path as well
 as the chunk size (in ha), can be set. Additionally, the verbosity can
-be set and the path to a log directory can be controlled. Might be
-extended by other options in the future.
+be set and the path to a log directory can be controlled. It also allow
+to set the number of retries and the delay between the successive
+retries when retrieving the resources. Might be extended by other
+options in the future.
 
 `get_resources()` data sets required for the calculation of indicators
 can be made available. The function supports the specification of
@@ -23,7 +25,7 @@ default), all available resources will be prepared.
 ## Usage
 
 ``` r
-mapme_options(..., outdir, chunk_size, retries, verbose, log_dir)
+mapme_options(..., outdir, chunk_size, retries, delay, verbose, log_dir)
 
 get_resources(x, ...)
 
@@ -57,8 +59,19 @@ prep_resources(
 
 - retries:
 
-  A numeric of length one indicating the number or re-tries the package
-  should attempt to make a resource available. Defaults to 3.
+  An integer of length one indicating the number of retries the package
+  should attempt to make a resource available. This commands the GDAL
+  option `GDAL_HTTP_MAX_RETRY` (see
+  [here](https://gdal.org/en/stable/user/configoptions.html#:~:text=GDAL_HTTP_MAX_RETRY)).
+  Defaults to 3.
+
+- delay:
+
+  An integer of length one indicating the number of seconds the package
+  should wait between the successive retries. This commands the GDAL
+  option `GDAL_HTTP_RETRY_DELAY` (see
+  [here](https://gdal.org/en/stable/user/configoptions.html#:~:text=GDAL_HTTP_RETRY_DELAY)).
+  Defaults to 10.
 
 - verbose:
 
@@ -68,9 +81,12 @@ prep_resources(
 
   A character path pointing toward a GDAL-writable destination used to
   log erroneous assets. Defaults to NULL, meaning that erroneous assets
-  will not be serialized to disk. If specified, a GPKG named
+  will not be serialized to disk. If specified, a GPKG file named
   `file.path(log_dir, paste0(Sys.Date(), "_mapme-error-assets.gpkg"))`
-  will be created and appended to in case of erroneous assets.
+  will be created and appended to in case of erroneous assets. In the
+  case some of the resources could not be downloaded, their list will be
+  written to a GPKG file named
+  `file.path(log_dir, paste0(Sys.Date(), "_", {resource_name}, "_mapme-error-resources.gpkg"))`.
 
 - x:
 
@@ -112,13 +128,16 @@ resources as `sf` and `SpatRaster`-objects.
 library(mapme.biodiversity)
 mapme_options()
 #> $outdir
-#> [1] "/tmp/RtmpcbUy3v/mapme-data"
+#> [1] "/tmp/RtmpRLzD8o/mapme-data"
 #> 
 #> $chunk_size
 #> [1] 1e+08
 #> 
 #> $retries
 #> [1] 3
+#> 
+#> $delay
+#> [1] 10
 #> 
 #> $verbose
 #> [1] FALSE
