### 2026-02-06 #9378 @jatkinson1000
Hi @goergen95 please could you give me an update over the next week on how you are progressing with the paper updates and if you still intend for this work to be considered for publication in JOSS.
