+++ paper/paper.bib
@@ -65,3 +65,19 @@
 	year = {2024},
 	note = {Publisher: EcoEvoRxiv},
 }
+
+@article{ramiandrisoa2026,
+    doi = {10.1371/journal.pone.0342093},
+    author = {Ramiandrisoa, Diamondra AND Bédécarrats, Florent AND Wong, Melvin H. L. AND Razanakoto, Thierry},
+    journal = {PLOS ONE},
+    publisher = {Public Library of Science},
+    title = {Impact of protected areas on deforestation in Madagascar from 2000 to 2023: A pre-analysis plan},
+    year = {2026},
+    month = {02},
+    volume = {21},
+    url = {https://doi.org/10.1371/journal.pone.0342093},
+    pages = {1-15},
+    abstract = {Protected areas are the most prevalent strategy to mitigate biodiversity loss and deforestation, especially in biodiversity hotspots like Madagascar. This pre-analysis plan outlines the data, methods, and identification strategy that will be used to assess the impact on deforestation of terrestrial protected areas created in Madagascar between 2002 and 2022. We will employ coarsened exact matching and difference-in-differences methods to evaluate forest cover loss, leveraging 24 years of high-resolution satellite data. We will incorporate buffer zones to assess spillover and leakage effects. This study addresses limitations of previous ones by combining accurate protected area delimitation and characteristics, a longer temporal coverage, improved characterization of the diverse forest ecosystems, and state-of-the art econometric methods. We will then assess heterogeneity of conservation effects, to better understand the determinants of protected area effectiveness.},
+    number = {2},
+
+}
