### 2026-02-11 #9378 @jatkinson1000
Thanks @goergen95 I'll get started on this.
In the meantime please can you check [`10.32614/r.manuals`](https://doi.org/10.32614/r.manuals) as being a DOI for R and update your references accordingly if that is correct.
