### 2026-02-11 #9378 @goergen95
@jatkinson1000 running the `citation()` command in R gives:
```
> citation()
To cite R in publications use:

  R Core Team (2025). _R: A Language and Environment for Statistical
  Computing_. R Foundation for Statistical Computing, Vienna, Austria.
  <https://www.R-project.org/>.

A BibTeX entry for LaTeX users is

  @Manual{,
    title = {R: A Language and Environment for Statistical Computing},
    author = {{R Core Team}},
    organization = {R Foundation for Statistical Computing},
    address = {Vienna, Austria},
    year = {2025},
    url = {https://www.R-project.org/},
  }

We have invested a lot of time and effort in creating R, please cite it
when using it for data analysis. See also ‘citation("pkgname")’ for
citing R packages.
```

which [rOpenSci](https://ropensci.org/blog/2021/11/16/how-to-cite-r-and-r-packages/#how-to-cite-r) also references on how to cite R. Skimming through the most similar JOSS publications listed above, only one of them ([glottospace](https://joss.theoj.org/papers/10.21105/joss.04303)) references R Core with a DOI but it seems obviously wrong (see [here](https://doi.org/10.1007/978-3-540-74686-7)). So in case you too are not aware of a canonical DOI to cite R Core in JOSS publications, I'd prefer not to add the one mentioned above which refers to a set of R manuals.
