+++ dev/reference/mapme.md
@@ -128,7 +128,7 @@ resources as `sf` and `SpatRaster`-objects.
 library(mapme.biodiversity)
 mapme_options()
 #> $outdir
-#> [1] "/tmp/RtmpL2NES5/mapme-data"
+#> [1] "/tmp/RtmpiE6N7Q/mapme-data"
 #> 
 #> $chunk_size
 #> [1] 1e+08
