### 2026-03-01 #9378 @jatkinson1000
Hi there @hydrogeohc, @kanishkan91, @samherniman, @jamesdamillington, and @ifoxfoot!

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is an R package for analysing biodiversity indicators and threats based on geodata-sources without the need for dedicated platforms.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Note that the scope of the journal [recently changed](https://github.com/openjournals/joss-reviews/issues/9378#issuecomment-3808150874), with a few more details being required in papers. Much of the process and sentiments remains as it always has, however.
