### 2026-03-06 #9378 @jatkinson1000
Hi there @hydrogeohc, @kanishkan91, @samherniman, @jamesdamillington, and @ifoxfoot, just bumping this request to review for JOSS. Please let me know if you are available to do this.
