### 2026-03-06 #9378 @kanishkan91
> Hi there [@hydrogeohc](https://github.com/hydrogeohc), [@kanishkan91](https://github.com/kanishkan91), [@samherniman](https://github.com/samherniman), [@jamesdamillington](https://github.com/jamesdamillington), and [@ifoxfoot](https://github.com/ifoxfoot), just bumping this request to review for JOSS. Please let me know if you are available to do this.

Hi @jatkinson1000 unfortunately, I'm a bit slammed with JOSS reviewing and editing, so will pass.
