### 2026-03-10 #9378 @jatkinson1000
Thanks @samherniman and @hydrogeohc I'll give it a couple of days to see if we can get a third reviewer (JOSS preference) but if not I'll start the review to avoid you needing to wait for too long.
