### 2026-03-14 0836196
docs: add joss status badge to README
