### 2026-03-24 #10166 @equitable-equations
I don't see community guidelines in the repo. I've already opened an issue, so other reviewers presumably don't need to duplicate.

**Update 2026-03-25**: resolved by developers, issue closed.
