+++ README.md
@@ -64,6 +64,8 @@ remotes::install_github("https://github.com/mapme-initiative/mapme.biodiversity"
 Below is a list of the resources currently supported by
 `mapme.biodiversity`.
 
+    ## mapme.biodiversity 0.9.5.9002
+
 | name | description | licence |
 |:---|:---|:---|
 | accessibility_2000 | Accessibility data for the year 2000 from the Global Accessibility Map project | See JRC data policy: <https://joint-research-centre.ec.europa.eu/jrc-mission-statement-work-programme/data-policy_en> |
@@ -164,6 +166,8 @@ library(mapme.biodiversity)
 library(sf)
 ```
 
+    ## Linking to GEOS 3.14.1, GDAL 3.12.2, PROJ 9.7.0; sf_use_s2() is TRUE
+
 Once you have decided on an indicator you are interested in, you can
 start by making the required resource available for your portfolio.
 Using `mapme_options()` you can set an output directory, control the
@@ -201,15 +205,7 @@ aoi <- system.file("extdata", "sierra_de_neiba_478140_2.gpkg", package = "mapme.
   ) %>%
   calc_indicators(calc_treecover_area_and_emissions(years = 2016:2017, min_size = 1, min_cover = 30)) %>%
   portfolio_long()
-```
 
-    ## Resource 'gfw_treecover' is already available.
-
-    ## Resource 'gfw_lossyear' is already available.
-
-    ## Resource 'gfw_emissions' is already available.
-
-``` r
 aoi
 ```
 
