### 2026-03-25 2955753
docs: adds contributing section to readme (#481)
