+++ README.md
@@ -1,15 +1,6 @@
-
 <!-- badges: start -->
 
-[![status](https://joss.theoj.org/papers/0d1960441cd29c1a250366db291c96c9/status.svg)](https://joss.theoj.org/papers/0d1960441cd29c1a250366db291c96c9)
-[![R-CMD-check](https://github.com/mapme-initiative/mapme.biodiversity/actions/workflows/R-CMD-check.yaml/badge.svg?branch=main)](https://github.com/mapme-initiative/mapme.biodiversity/actions)
-[![Coverage
-Status](https://img.shields.io/codecov/c/github/mapme-initiative/mapme.biodiversity/master.svg)](https://app.codecov.io/github/mapme-initiative/mapme.biodiversity?branch=main)
-[![CRAN
-status](https://badges.cranchecks.info/worst/mapme.biodiversity.svg)](https://cran.r-project.org/web/checks/check_results_mapme.biodiversity.html)
-[![CRAN
-version](https://www.r-pkg.org/badges/version/mapme.biodiversity)](https://CRAN.R-project.org/package=mapme.biodiversity)
-[![License](https://img.shields.io/badge/License-GPL%20(%3E=3)-brightgreen.svg?style=flat)](https://choosealicense.com/licenses/gpl-3.0/)
+[![status](https://joss.theoj.org/papers/0d1960441cd29c1a250366db291c96c9/status.svg)](https://joss.theoj.org/papers/0d1960441cd29c1a250366db291c96c9) [![R-CMD-check](https://github.com/mapme-initiative/mapme.biodiversity/actions/workflows/R-CMD-check.yaml/badge.svg?branch=main)](https://github.com/mapme-initiative/mapme.biodiversity/actions) [![Coverage Status](https://img.shields.io/codecov/c/github/mapme-initiative/mapme.biodiversity/master.svg)](https://app.codecov.io/github/mapme-initiative/mapme.biodiversity?branch=main) [![CRAN status](https://badges.cranchecks.info/worst/mapme.biodiversity.svg)](https://cran.r-project.org/web/checks/check_results_mapme.biodiversity.html) [![CRAN version](https://www.r-pkg.org/badges/version/mapme.biodiversity)](https://CRAN.R-project.org/package=mapme.biodiversity) [![License](https://img.shields.io/badge/License-GPL%20(%3E=3)-brightgreen.svg?style=flat)](https://choosealicense.com/licenses/gpl-3.0/)
 
 <!-- badges: end -->
 
@@ -17,26 +8,9 @@ version](https://www.r-pkg.org/badges/version/mapme.biodiversity)](https://CRAN.
 
 ## About
 
-Biodiversity areas, especially primary forests, provide multiple
-ecosystem services for the local population and the planet as a whole.
-The rapid expansion of human land use into natural ecosystems and the
-impacts of the global climate crisis put natural ecosystems and the
-global biodiversity under threat.
+Biodiversity areas, especially primary forests, provide multiple ecosystem services for the local population and the planet as a whole. The rapid expansion of human land use into natural ecosystems and the impacts of the global climate crisis put natural ecosystems and the global biodiversity under threat.
 
-The mapme.biodiversity package helps to analyse a number of biodiversity
-related indicators and biodiversity threats based on freely available
-geodata-sources such as the Global Forest Watch. It supports
-computational efficient routines and heavy parallel computing in
-cloud-infrastructures such as AWS or Microsoft Azure using in the
-statistical programming language R. The package allows for the analysis
-of global biodiversity portfolios with a thousand or millions of AOIs
-which is normally only possible on dedicated platforms such as the
-Google Earth Engine. It provides the possibility to e.g. analyse the
-World Database of Protected Areas (WDPA) for a number of relevant
[…]
