### 2026-03-25 #10166 @equitable-equations
@jatkinson1000 how closely should we be assessing functionality? The examples and unit tests run (as already checked by CRAN), and I've directly confirmed that the code in the paper and repo readme work as described.
