### 2026-03-25 #10166 @jatkinson1000
@equitable-equations thanks for asking

The JOSS review criteria states:
> Reviewers are expected to install the software they are reviewing and to verify the core functionality of the software.

This isn't expected to be exhaustive, but should test enough of the functionality to confirm that it does what it describes in the paper and give confidence that any other features will also work as described. It should also not require delving into the code (unless you want to) but is instead focussed on using the software as described in documentation.

The ideal situation is that the authors provide tutorial(s) or example(s) accompanying the code for new users to familiarise themselves. If there isn't one then it may make sense to suggest adding one. Alternatively, if it is more of a workflow-based package then they might point you at some open-source data and you can piece together a small workflow to do something as described in the documentation.

The real question we are trying to answer here is whether you feel satisfied that it does what it claims to in the paper, and whether you feel like, as a result of your review, you could take this package and make use of it in another situation to tackle something.

Hope this helps, but please ask further if needed.
