### 2026-03-25 #10166 @equitable-equations
Thanks, @jatkinson1000 this makes sense and is in line with my previous understanding. I anticipate completing my checklist this weekend.
