### 2026-03-27 #10166 @jatkinson1000
Hi @samherniman please could you give us an update on the state of your review?
I see that the checklist is generated but nothing is checked off yet.

If you could please find some time to work through this and raise any feedback for the authors over the next 2 weeks that would be appreciated, thank you. JOSS tries to aim for a 4-week turnaround in reviews, including time for authors to address feedback.
