### 2026-03-27 #10166 @jatkinson1000
Hi @hydrogeohc please could you give us an update on the state of your review?
You haven't generated your checklist yet using `@editorialbot generate my checklist`.

If you could please find some time to start this and raise any feedback for the authors over the next 2 weeks that would be appreciated, thank you. JOSS tries to aim for a 4-week turnaround in reviews, including time for authors to address feedback.
