### 2026-03-31 #10166 @equitable-equations
@jatkinson1000 my checklist is complete and the paper has my blessing.
