### 2026-04-02 #10166 @jatkinson1000
Great, thank you @equitable-equations

@samherniman I see you are yet to start your review. Please could you give us an update on how things are going and when you might be able to work through it to provide feedback to the author?

Similarly @hydrogeohc please could you generate your checklist and give us an update on progress?

Ideally I'd like a first pass completed by the end of next week so that we can keep on track and move this forwards please, though do let me and the author know if you are going to need longer.
