### 2026-04-10 #10166 @jatkinson1000
Thanks @samherniman - I see you have not checked off "Reproducibility" on your list.
Is this just an oversight, or do you need the authors to address anything for this to be considered passing?

@hydrogeohc I see you are yet to generate a checklist - are you still available and willing to review this paper?
Please let me know before early next week so I can move this forward.
