### 2026-04-23 #10166 @goergen95
Thanks @jatkinson1000 and thank you to the reviewers for the positive evaluation of our submission!

We are planning to integrate the paper into the default branch of our repository (PR linked above). Once this is done I would update the branch to be used to render the paper in this review thread. Is this ok with the JOSS processes or does it actually interfere at the current stage?
