### 2026-04-23 d04cdbf
update worldclim test data
