+++ tests/testthat/_snaps/calc_precipitation_wc.md
@@ -3,6 +3,6 @@
     Code
       result_exact$value
     Output
-       [1]  26.36556  25.85101  66.69062  82.51883 330.19525  60.50434  97.28488
-       [8] 150.38458 217.67076 160.12506  69.11417  24.80346
+       [1]  26.36556  25.85101  66.69062  82.51883 330.19525  75.18739  97.28488
+       [8] 105.82425 217.67076 160.12506  59.46145  24.80346
 
