### 2026-04-23 #10166 @equitable-equations
Congratulations @goergen95 this is a cool project and a fun repo to review.
