+++ R/register.R
@@ -4,6 +4,11 @@
 .pkgenv$indicators <- list()
 .pkgenv$avail_resources <- list()
 
+.onAttach <- function(libname, pkgname) {
+  packageStartupMessage(paste(pkgname, utils::packageVersion(pkgname)))
+  invisible()
+}
+
 .onLoad <- function(libname, pkgname) {
   outdir <- tempfile(pattern = "mapme-data-")
   dir.create(outdir, showWarnings = FALSE)
@@ -16,8 +21,6 @@
     verbose = TRUE
   )
   .check_system_requirements()
-  packageStartupMessage(paste(pkgname, utils::packageVersion(pkgname)))
-  invisible()
 }
 
 .probe_dsn <- function(dsn) {
