+++ paper/paper.bib
@@ -1,83 +1,0 @@
-@misc{rcore,
-  title = {{R: A language and environment for statistical computing}},
-  author = {{R Core Team}},
-  organization = {R Foundation for Statistical Computing},
-  address = {Vienna, Austria},
-  year = {2022},
-  url = {https://www.R-project.org/},
-}
-
-@misc{mapme,
-  title = {mapme.biodiversity: Efficient Monitoring of Global Biodiversity Portfolios},
-  author = {Darius A. Görgen and Om Prakash Bhandari},
-  year = {2025},
-  note = {R package version 0.9.4},
-  url = {https://mapme-initiative.github.io/mapme.biodiversity/},
-  doi = {10.32614/CRAN.package.mapme.biodiversity}
-}
-
-
-@article{juffe-bignoli_delivering_2024,
-	title = {Delivering {Systematic} and {Repeatable} {Area}-{Based} {Conservation} {Assessments}: {From} {Global} to {Local} {Scales}},
-	volume = {13},
-	copyright = {http://creativecommons.org/licenses/by/3.0/},
-	issn = {2073-445X},
-	shorttitle = {Delivering {Systematic} and {Repeatable} {Area}-{Based} {Conservation} {Assessments}},
-	url = {https://www.mdpi.com/2073-445X/13/9/1506},
-	doi = {10.3390/land13091506},
-	abstract = {Protected and conserved areas are a key area-based strategy to address the biodiversity and climate change crises. Indicators are fundamental to understanding performance over time. The Digital Observatory for Protected Areas (DOPA) was born in 2013 as a set of open-access web services and applications to be used to assess, monitor, and report on protected and conserved areas. For over a decade, it has delivered over 50 indicators to support policy processes, national and regional governments, and practitioners. DOPA has also developed a versatile and efficient back-end approach that is widely applicable in other area-based conservation contexts. Here, we describe the methods and workflows behind DOPA’s back end and provide examples of policy relevant questions it can answer. We discuss the key advantages and limitations of this systematic and replicable approach and explore the use of this back-end architecture to inform progress in area-based conservation targets for the following decades. This approach, embedded in multiple services provided by the Knowledge Centre for Biodiversity of the European Commission (KCBD), can also support the implementation and monitoring of area-based targets of the Kunming Montreal Global Biodiversity Framework at international, regional, and national levels.},
-	language = {en},
-	number = {9},
-	urldate = {2025-10-08},
-	journal = {Land},
-	author = {Juffe-Bignoli, Diego and Mandrici, Andrea and Delli, Giacomo and Niamir, Aidin and Dubois, Grégoire},
-	month = sep,
-	year = {2024},
-	note = {Publisher: Multidisciplinary Digital Publishing Institute},
-	keywords = {biodiversity, protected areas indicators, spatial planning},
-	pages = {1506},
-}
-  
-@article{hansen_high-resolution_2013,
-	title = {High-resolution global maps of 21st-century forest cover change},
-	volume = {342},
-	doi = {10.1126/science.1244693},
-	url = {http://www.sciencemag.org/content/342/6160/850.short},
-	number = {6160},
-	urldate = {2016-01-13},
-	journal = {science},
-	author = {Hansen, Matthew C. and Potapov, Peter V. and Moore, Rebecca and Hancher, Matt and Turubanova, S. A. and Tyukavina, Alexandra and Thau, David and Stehman, S. V. and Goetz, S. J. and Loveland, T. R. and {others}},
-	year = {2013},
-	pages = {850--853},
-}
-
-@article{griffith_bon_2024,
[…]
