+++ paper/paper.md
@@ -1,218 +1,0 @@
----
-title: "mapme.biodiversity: Efficient Monitoring of Global Biodiversity Portfolios"
-tags:
-  - R
-  - conservation research
-  - protected areas
-  - spatial analysis
-authors:
-  - name: Darius A. Görgen
-    orcid: 0009-0008-5503-7704
-    affiliation: 1
-  - name: Om Prakash Bhandari
-    affiliation: 2
-  - name: Andreas Petutschnig
-    orcid: 0000-0001-5029-2425
-    affiliation: 3
-  - name: Zivan Karaman
-    orcid: 0000-0002-8933-4589
-    affiliation: 4
-  - name: Florent Bédécarrats
-    orcid: 0000-0003-1001-5540
-    affiliation: 5
-  - name: Johannes Schielein
-    orcid: 0000-0001-6286-0688
-    affiliation: 6
-affiliations:
- - name: Department of Landscape Ecology, University of Münster, Germany
-   index: 1
- - name: Department of Geodetic Engineering, University of Bonn, Germany
-   index: 2
- - name: adesso SE, Dortmund, Germany
-   index: 3
- - name: Independent consultant, France
-   index: 4
- - name: UMI SOURCE, Université Paris-Saclay, UVSQ, IRD, Guyancourt, France
-   index: 5
- - name: KfW Development Bank, Bonn, Germany
-   index: 6
-date: 03 November 2025
-bibliography: paper.bib
-software_repository_url: www.github.com/mapme-initiative/mapme.biodiversity
----
-
-# Summary
-
-The \texttt{mapme.biodiversity} \texttt{R} package provides an open and
-reproducible framework for computing biodiversity-related indicators
-from multiple global geospatial datasets. It streamlines access,
-preprocessing, and analysis of spatial data to produce standardized,
-area-based metrics that can be used for conservation monitoring, policy
-evaluation, and research. By enabling reproducible, scalable, and extendable 
-analyses across thematic datasets, \texttt{mapme.biodiversity} helps
-to bridge the gap between raw geospatial information and decision-relevant
-biodiversity indicators.
-
-# Statement of need
-
-To prevent biodiversity loss at scale, conservation researchers and
-practitioners require area-based indicators derived from diverse spatial
-datasets. However, relevant data sources are dispersed across multiple 
-repositories and platforms, each with distinct access protocols, 
-formats, and documentation standards.
-
-Most existing tools focus on a specific data source or domain, offering
-limited interoperability and requiring users to learn multiple
-interfaces. This fragmentation imposes a high cognitive and technical
-burden, especially on users who are not remote sensing specialists but
-need spatial data for research, monitoring, or policy evaluation.
-
-The \texttt{mapme.biodiversity} \texttt{R} package provides a unified
-interface to access and process a wide range of spatial datasets
-relevant to conservation and environmental management. It enables users
-to derive standardized, area-based indicators at scale, supports
-reproducible workflows in \texttt{R}, and facilitates integration with
-other tools via export to standard spatial formats.
-
-# State of the field
-
-The Digital Observatory for Protected Areas provides a centralized, 
-server-based system developed by the European Commission to compute global
-protected-area indicators through standardized, automated workflows
-[@juffe-bignoli_delivering_2024]. In contrast, \texttt{mapme.biodiversity} offers
-an R-native, decentralized framework that enables users to reproduce
-similar area-based analyses locally, adapt them to specific contexts,
-and extend them with additional datasets.
-
-The Global Forest Watch (\texttt{GFW}) API, developed by the World 
-Resource Institute, provides access to selected global forest monitoring 
-datasets such as tree cover, loss, gain, biomass, and fire activity 
[…]
