+++ paper/README.md
@@ -1,15 +1,0 @@
-We use the [inara](https://github.com/openjournals/inara) docker
-image to render the article. Please make sure that docker is 
-installed on your system and can be run without sudo.
-
-To render the JOSS article in draft mode:
-
-```
-make draft
-```
-
-and for production mode:
-
-```
-make article
-```
