### 2026-04-30 #10166 @goergen95
@jatkinson1000: We released version `0.9.6` with the feedback from the review included. The DOI for this release is: `10.5281/zenodo.19908654`.
