+++ llms.txt
@@ -32,6 +32,7 @@ KfW.
 The package and its dependencies can be installed from CRAN via:
 
 ``` r
+
 install.packages("mapme.biodiversity", dependencies = TRUE)
 ```
 
@@ -42,6 +43,7 @@ Windows and macOS binary packages are available from here.
 To install the development version, use the following command:
 
 ``` r
+
 if (isFALSE(require("remotes", quietly = TRUE))) install.packages("remotes", dependencies = TRUE)
 remotes::install_github("https://github.com/mapme-initiative/mapme.biodiversity", dependencies = TRUE)
 ```
@@ -55,43 +57,43 @@ Below is a list of the resources currently supported by
 ## mapme.biodiversity 0.9.6
 ```
 
-| name                             | description                                                                                                                | licence                                                                                                                    |
-|:---------------------------------|:---------------------------------------------------------------------------------------------------------------------------|:---------------------------------------------------------------------------------------------------------------------------|
-| accessibility_2000               | Accessibility data for the year 2000 from the Global Accessibility Map project                                             | See JRC data policy: <https://joint-research-centre.ec.europa.eu/about-us/service-standards-and-principles/data-policy_en> |
-| acled                            | Armed Conflict Location & Event Data (ACLED)                                                                               | Visit acleddata.com                                                                                                        |
-| biodiversity_intactness_index    | Biodiversity Intactness Index                                                                                              | CC-BY-4.0                                                                                                                  |
-| chelsa                           | Climatologies at High resolution for the Earth Land Surface Areas (CHELSA)                                                 | Unknown - Must cite!                                                                                                       |
-| chirps                           | Climate Hazards Group InfraRed Precipitation with Station data (CHIRPS)                                                    | CC - unknown                                                                                                               |
-| esalandcover                     | Copernicus Land Monitoring Service (CLMS) 100 meter land cover product                                                     | CC-BY 4.0                                                                                                                  |
-| fritz_et_al                      | Drivers of deforestation in the tropics                                                                                    | CC-BY 4.0                                                                                                                  |
-| gfw_emissions                    | Global Forest Watch - CO2 Emssions caused by forest cover loss                                                             | CC-BY 4.0                                                                                                                  |
-| gfw_lossyear                     | Global Forest Watch - Year of forest cover loss occurrence                                                                 | CC-BY 4.0                                                                                                                  |
-| gfw_treecover                    | Global Forest Watch - Percentage of canopy closure in 2000                                                                 | CC-BY 4.0                                                                                                                  |
-| global_surface_water_change      | Global Surface Water - Change of water occurrence intensity                                                                | <https://www.copernicus.eu/en/access-data>                                                                                 |
-| global_surface_water_occurrence  | Global Surface Water - Percentage of water occurrence                                                                      | <https://www.copernicus.eu/en/access-data>                                                                                 |
-| global_surface_water_recurrence  | Global Surface Water - Percentage of water recurrence                                                                      | <https://www.copernicus.eu/en/access-data>                                                                                 |
[…]
