### 2026-05-03 #10166 @jatkinson1000
@goergen95 please can you update the Zenodo entry to add affiliation for each of the authors, ensuring that they are verbatim consistent with the paper.
