### 2026-05-03 #10166 @jatkinson1000
Hi @goergen95

In addition to the above, I have now read the paper and echo the comments from the reviewers that this looks to be a useful quality piece of software and a well-written paper. I have no notable changes to propose from the editorial side.

My only comment is that the formatting of the Authors in the reference for "Ramiandrisoa, F. A. W., Diamondra AND Bédécarrats. (2026)" looks to be wrong. Please check the BibTeX and change the "AND" to "and".

Once you have done this let me know what you want to do with respect to archiving and release and I'll then set these accordingly.
