+++ reference/burned_area.md
@@ -57,13 +57,28 @@ aoi <- system.file("extdata", "sierra_de_neiba_478140_2.gpkg",
   get_resources(get_mcd64a1(years = 2010)) %>%
   calc_indicators(calc_burned_area(engine = "extract")) %>%
   portfolio_long()
-#> Error : HTTP content type response 'text/plain' not defined for this operation.
-#> Error in (function (x, name = "mcd64a1", type = "raster", outdir = mapme_options()[["outdir"]],  : 
-#>   Download for MCD64A1 resource was unsuccesfull
-#> Error in .check_footprints(resource, resource_name): Download for resource mcd64a1 failed.
-#> Returning unmodified portfolio.
 
 aoi
-#> Error: object 'aoi' not found
+#> Simple feature collection with 12 features and 8 fields
+#> Geometry type: POLYGON
+#> Dimension:     XY
+#> Bounding box:  xmin: -71.80933 ymin: 18.57668 xmax: -71.33201 ymax: 18.69931
+#> Geodetic CRS:  WGS 84
+#> # A tibble: 12 × 9
+#>    WDPAID ISO3  assetid indicator   datetime            variable    unit  value
+#>     <dbl> <chr>   <int> <chr>       <dttm>              <chr>       <chr> <dbl>
+#>  1 478140 DOM         1 burned_area 2010-12-01 00:00:00 burned_area ha      0  
+#>  2 478140 DOM         1 burned_area 2010-11-01 00:00:00 burned_area ha      0  
+#>  3 478140 DOM         1 burned_area 2010-10-01 00:00:00 burned_area ha      0  
+#>  4 478140 DOM         1 burned_area 2010-09-01 00:00:00 burned_area ha      0  
+#>  5 478140 DOM         1 burned_area 2010-08-01 00:00:00 burned_area ha      0  
+#>  6 478140 DOM         1 burned_area 2010-07-01 00:00:00 burned_area ha      0  
+#>  7 478140 DOM         1 burned_area 2010-06-01 00:00:00 burned_area ha      0  
+#>  8 478140 DOM         1 burned_area 2010-05-01 00:00:00 burned_area ha      0  
+#>  9 478140 DOM         1 burned_area 2010-04-01 00:00:00 burned_area ha      0  
+#> 10 478140 DOM         1 burned_area 2010-03-01 00:00:00 burned_area ha     42.8
+#> 11 478140 DOM         1 burned_area 2010-02-01 00:00:00 burned_area ha      0  
+#> 12 478140 DOM         1 burned_area 2010-01-01 00:00:00 burned_area ha      0  
+#> # ℹ 1 more variable: geom <POLYGON [°]>
 # }
 ```
