+++ reference/mapme.md
@@ -128,7 +128,7 @@ resources as `sf` and `SpatRaster`-objects.
 library(mapme.biodiversity)
 mapme_options()
 #> $outdir
-#> [1] "/tmp/RtmpO4Q5SH/mapme-data"
+#> [1] "/tmp/RtmpKMzxtu/mapme-data"
 #> 
 #> $chunk_size
 #> [1] 1e+08
